+++ crest/Assets/Crest/Crest/Scripts/WaterBody.cs
@@ -198,6 +198,8 @@ namespace Crest
     {
         public bool Validate(OceanRenderer ocean, ValidatedHelper.ShowMessage showMessage)
         {
+            var isValid = true;
+
             // This will also return disabled objects. Safe to use in this case.
             if (Resources.FindObjectsOfTypeAll<OceanRenderer>().Length == 0)
             {
@@ -208,7 +210,7 @@ namespace Crest
                     ValidatedHelper.MessageType.Error, this
                 );
 
-                return false;
+                isValid = false;
             }
 
             if (Mathf.Abs(transform.lossyScale.x) < 2f && Mathf.Abs(transform.lossyScale.z) < 2f)
@@ -220,20 +222,12 @@ namespace Crest
                     ValidatedHelper.MessageType.Error, this
                 );
 
-                return false;
+                isValid = false;
             }
 
-            if (transform.eulerAngles.magnitude > 0.0001f)
-            {
-                showMessage
-                (
-                    $"There must be no rotation on the water body GameObject, and no rotation on any parent. Currently the rotation Euler angles are {transform.eulerAngles}.",
-                    "Reset the rotations on this GameObject and all parents to 0.",
-                    ValidatedHelper.MessageType.Error, this
-                );
-            }
+            isValid = isValid && ValidatedHelper.ValidateNoRotation(this, transform, showMessage);
 
-            return true;
+            return isValid;
         }
     }